ResourceRoutesCOntroller moved to src\Http directory
